requirements.txt
Django==1.9
django-filter==0.15.3
djangorestframework==3.5.3
Markdown==2.6.7
Pillow==3.4.2
